server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Andrii Momotenko <momotenko536@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.3"}
uvicorn = "^0.27.0.post1"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
